refactor: extract CSRF security logic into useCSRFSecurity hook
- Created new custom hook useCSRFSecurity for token management
- Reduced Contact component from 733 to 574 lines (159 line reduction)
- Improved separation of concerns and reusability
- All security logic now centralized in dedicated hook
- Added proper TypeScript interfaces and return types
- Maintains all existing functionality while improving code organization
